+++ webpack.config.js
@@ -2,14 +2,23 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 const path = require('path');
 
 module.exports = {
-	entry: './src/index.js',
+	entry: {
+		index: './src/index.js',
+		'context-action': './src/context-action.js',
+		'modal-viewer': './src/modal-viewer.js',
+		hub: './src/hub.js'
+	},
 	output: {
 		path: path.resolve(__dirname, './dist'),
-		filename: 'index_bundle.js',
+		filename: '[name].js',
 	},
-	plugins: [new HtmlWebpackPlugin({
-		template: 'index.html'
-	})],
+	plugins: [
+		new HtmlWebpackPlugin({
+			template: 'index.html',
+			filename: 'index.html',
+			chunks: ['index']
+		})
+	],
 	module: {
 		rules: [
 			{